파일 분리 및 media query style 추가
반응형 디자인을 적용하기 위해 media query 함수를 파일로 분리하여
추가한다.
