+++ src/styles/Container.jsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+const Container = styled.div({
+  margin: '0 auto',
+  width: '90%',
+});
+
+export default Container;
